requirements.txt
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
TikTokLive==5.0.7
tenacity==8.2.3
python-dateutil==2.8.2
pydantic==2.5.2
colorlog==6.7.0
aiohttp==3.9.1
asyncio==3.4.3
pytest==7.4.3
pytest-asyncio==0.21.1
mypy==1.7.1
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
fastapi-utils==0.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
plotly==5.18.0
pandas==2.1.3 